Trying to remove LOI in the ore heater.
Problem with this is that now the ore composition does not match between
what is specified in the system vars and the actual ore. Causes issues
when we call iron_species_from_reduction_degree.

Actually, this problem might have always been here in the DRI-EAF system,
but was covered up by the injected O2. probs need another function to
get the ore composition from the actual ore or dri object, rather than
relying on the system_vars. MAJOR BUG!